Unify common code for StarOffice format
